tests: the two new cases spell a 64-bit type, not `unsigned long`
Both failed on the Windows lanes and passed everywhere else. `unsigned
long` is 64 bits under LP64 and 32 under LLP64, and both cases meant 64.

`control_debug_segment_mov_inline_asm_x64` parses with the host's
default target and emits for LinuxX64, so off an LP64 host the operand
came out 32-bit and the control-register rule rejected it -- correctly.
The fixture shifted a 32-bit value by 62 and read back the wrong length.

+++ src/c5/tests/codegen.rs
@@ -4704,7 +4704,7 @@ fn control_debug_segment_mov_inline_asm_x64() {
     //   read  cr/dr -> gpr : 0F 20 / 0F 21 ; write gpr -> cr/dr : 0F 22 / 0F 23
     //   read  seg   -> gpr : 8C.
     let program = super::compile_str_bare(
-        "typedef unsigned long UN;\n\
+        "typedef unsigned long long UN;\n\
          UN rcr0(void){ UN d; __asm__ __volatile__(\"mov  %%cr0,%0\":\"=r\"(d)); return d; }\n\
          UN rcr3(void){ UN d; __asm__ __volatile__(\"mov  %%cr3,  %0\":\"=r\"(d)); return d; }\n\
          void wcr0(UN v){ __asm__ __volatile__(\"mov  %0, %%cr0\"::\"r\"(v)); }\n\
